speed up

+++ my-everthing/src/slicing/graph_builder.py
@@ -1,13 +1,10 @@
 import os
 import networkx as nx
-from icecream import ic
 import config
 import logging
 import subprocess
 
 logger = logging.getLogger(__name__)
-
-ic.configureOutput(includeContext=True)
 
 class GraphBuilder:
     def __init__(self):
@@ -15,27 +12,35 @@ class GraphBuilder:
         pass
 
     def joern_parse(self, source_path, bin_file = "tmp.bin"):
+        logger.info(f"Start parsing {source_path} to {bin_file}")
         cmd = f'joern-parse  {source_path} -o {bin_file}'
         self.execute_command(cmd)
+        logger.info(f"Parse {source_path} to {bin_file}")
 
-    def joern_dump_graph(self, graph_dir="out_tmp", bin_file="tmp.bin",  graph_type='cpg'):
+    def joern_dump_graph(self, graph_dir="out_tmp", bin_file="tmp.bin",  graph_type='cpg', overwrite=config.GRAPH_OVERWRITE):
+        logger.info(f"Start dumping graph to {graph_dir}")
+        if os.path.exists(graph_dir) and not overwrite:
+            return
         if os.path.exists(graph_dir):
             os.system(f'rm -rf {graph_dir}')
         cmd = f'joern-export --repr {graph_type} {bin_file} --out {graph_dir}'
         self.execute_command(cmd)
+        logger.info(f"Dump graph to {graph_dir}")
     
     def execute_command(self, cmd):
-        logger.info("[GraphBuilder] Executing command: %s", cmd)
+        logger.info("Executing command: %s", cmd)
         try:
             result = subprocess.run(cmd, shell=True, check=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
-            logger.info("[GraphBuilder] Command output: %s", result.stdout)
+            logger.info("Command output: %s", result.stdout)
             if result.stderr:
-                logger.error("[GraphBuilder] Command error: %s", result.stderr)
+                logger.error("Command error: %s", result.stderr)
         except subprocess.CalledProcessError as e:
-            logger.error("[GraphBuilder] Command failed with error: %s", e.stderr)
+            logger.error("Command failed with error: %s", e.stderr)
     
     def read_graph(self, graph_path):
+        logger.info(f"Start reading graph {graph_path}")
         G = nx.drawing.nx_agraph.read_dot(graph_path)
+        logger.info(f"Read graph {graph_path}")
         return G
     
 
@@ -48,13 +53,13 @@ class GraphBuilder:
             G = self.read_graph(graph_filepath)
             G_combine.add_nodes_from(G.nodes(data=True))
             G_combine.add_edges_from(G.edges(data=True))
-            # ic(G_combine.number_of_nodes(), G_combine.number_of_edges())
-            logger.info(f"[GraphBuilder] Combine graph {filename} to G_combine")
+            logger.info(f"Combine graph {filename} to G_combine")
         return G_combine
     
     def save_graph(self, save_path, G):
+        logger.info(f"Start saving graph...")
         nx.drawing.nx_agraph.write_dot(G, save_path)
-        ic(save_path)
+        logger.info(f"Save graph to {save_path}")
 
 
 def test():
